+++ lib/catan/helpers/score_calculator.rb
@@ -1,17 +1,22 @@
 # frozen_string_literal: true
 
 class ScoreCalculator
-  def initialize(settlements:, cards:)
+  def initialize(settlements:, cities:, cards:)
     @settlements = settlements
+    @cities = cities
     @cards = cards
   end
 
   def calc(player)
-    settlements_count(player) + victory_cards_count(player)
+    settlements_count(player) + (2 * cities_count(player)) + victory_cards_count(player)
   end
 
   def settlements_count(player)
     @settlements.select { |settlement| settlement.owner.equal?(player) }.count
+  end
+
+  def cities_count(player)
+    @cities.select { |city| city.owner.equal?(player) }.count
   end
 
   def victory_cards_count(player)